Add relative path support to SSH source file mapping.
The SSH source file mapping now utilizes the CWD of the remote machine
to determine the path kind that is being used when the supplied path
mapping is a relative path.  This assumes that the CWD is an absolute
path (which it should be) and is used as a backup when the path in the
map cannot be determined to be an absolute path.